backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
feedparser
requests
beautifulsoup4
lxml
openai
python-dotenv
httpx
tenacity
